feat：完善显示

+++ modulesCore/homeScreen/src/main/java/cn/pivotstudio/modulec/homescreen/ui/adapter/ForestHoleAdapter.kt
@@ -70,7 +70,7 @@ class ForestHoleAdapter(
                 }
 
                 layoutItemForestHole.setOnClickListener {
-                    _context.navToSpecificHole(forestHole.holeId.toInt())
+                    _context.navToSpecificHole(forestHole.holeId)
                 }
 
                 layoutItemForestMoreList.setOnClickListener {
